requirements.txt
fastapi
uvicorn
requests
faiss-cpu
langchain